[Refactor] Moves functions pertaining utility to *Utils.java
